+++ components/index.ts
@@ -1,4 +1,4 @@
 export * from './Header';
-export * from './LanguageSwitch';
+export * from './LanguageSwitch2';
 export * from './SnackbarMessage';
 export * from './Layout';
